+++ src/model/items/Solid.cpp
@@ -4,12 +4,12 @@
 
 Solid::Solid() {}
 
-Solid::Solid(std::array<int, 2> pos)
+Solid::Solid(std::array<float, 2> pos)
 {
     position = pos;
 }
 
-Solid::Solid(int pos_x, int pos_y)
+Solid::Solid(float pos_x, float pos_y)
 {
     position = {pos_x, pos_y};
 }
@@ -21,12 +21,12 @@ Solid::Solid(const Solid &other)
 
 Solid::~Solid() {}
 
-void Solid::set_position(std::array<int, 2> new_pos)
+void Solid::set_position(std::array<float, 2> new_pos)
 {
     position = new_pos;
 }
 
-std::array<int, 2> Solid::get_position()
+std::array<float, 2> Solid::get_position()
 {
     return position;
 }
